embeddings_library/requirements.txt
PyYaml
jsonschema
scikit-learn
PyPDF2
faiss-cpu
